ALN/ingest_service/requirements.txt
python3-fastapi
python3-uvicorn[standard]
